hw1/requirements.txt
shiny>=0.8.1
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
